[#2033] [MOB-76] Move Currency Conversion from Advanced Settings to Settings
Closes #2033
